requirements.txt
openml
deap
update_checker
tqdm
stopit
xgboost
scikit-mdr
skrebate
tensorflow==2.4.1
scipy==1.5.4
tpot
autokeras
Cython
numpy==1.19.5
pipelineprofiler
auto-sklearn
torch
h2o
emcee
scikit-optimize
pyDOE
autogluon
Pillow==8.1
psutil==5.7.0
portalocker==2.0.0
termcolor
mxnet